Move FormListDownloader to management package
